requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
pinecone
pinecone-text
python-dotenv
openai
PyPDF2
python-docx
python-pptx
Pillow
pytesseract
pandas
numpy
tqdm
python-multipart
streamlit
sentence-transformers
flask
flask-cors
werkzeug
langgraph
tavily-python
requests